+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 0.2.7 <date>
+- Added command parser: `rpm -qa`.
+- Added command parser: `yum list installed`
+- Added command parser: `snap list`.
+- Changed `dpkg.py` to `packages.py` to have all packages parsers in one place.
+
 ### 0.2.6 08.03.2025
 
 - Added logfile parser: `/var/log/udev`.